Logging handler added to the Flask application. It can be accessed by views, by calling current_app.logger.
Unit tests for both the views and end points are passing for the
functional test of Job Epic, for the first stage, involving a
user creating a library (keep in mind, no checks have been
included in this step, e.g., can this user access this library).
This *should* come from other storyboard examples.

Utils added that creates logs and grabs the post data from a
request object. This was taken from the accounts service which is
an core component of the ADSWS.