Some large changes in this commit:
Alembic migration tools for use with SQLAlchemy and the PostgreSQL
database have been included. This uses the Flask-Migrate extension.
Currently, given the tests are creating and destroying the content
of the database, it is not needed.

End point to add and retrieve bibcodes for a given library have been
included. All tests are currently passing.

Travis building file included - could have minor errors, will have
to wait until this commit is made.

Coveralls added to the travis file for code coverage.