Large modifications made to include the Deletion Epic:
A DELETE end point was added to the /user/<>/libraries/<> url.

Unit and functional tests were updated to include this ability and test it.

In addition, tests were added such that if you DELETE a library that does
not exist, it will raise a NoResultFound exception from SQLAlchemy, which
is passed by the web services as a 410. This error code was chosen for the
defition given by the w3:

The requested resource is no longer available at the server and no forwarding
 address is known. This condition is expected to be considered permanent.
Clients with link editing capabilities SHOULD delete references to the Request-URI
after user approval. If the server does not know, or has no facility to
determine, whether or not the condition is permanent, the status code 404
(Not Found) SHOULD be used instead. This response is cacheable unless
indicated otherwise.

A 404 is not appropriate given the server does know that this content does
not exist in the database. Of course this does not apply if people are
attempting to retrieve never created before library IDs, but I do not currently
see a use case for this.

Finally, tests were included such that a user cannot have two libraries with
the same name. SQLAlchemy has this field set to unique, and the web server
passes on the IntegrityError as a 409 error:

The request could not be completed due to a conflict with the current state
of the resource. This code is only allowed in situations where it is expected
that the user might be able to resolve the conflict and resubmit the request.
The response body SHOULD include enough

Given the user CAN resubmit the request successfully if they change the name,
this error code seems appropraite. An error message is supplied  via the web
server explaining that you cannot have duplicated library names.