Changes were made to address issue #1 and #3
The endpoints were changed so that they refer to only the UUID of
the library.

Class included so that PostgreSQL does not need to be used if not
wanted in the future. The class allows UUIDs.

The header from ADSWS will contain the api user ID.

Tests were updated to no longer pass a user ID in the URL.