Started to implement the add/remove of permissions for users and libraries.
test_webservices should currently be broken and waiting to be fixed
in standard TDD style.

note: the permissions broke the non-existent library return. More
specifically - if a user deletes a library, there are no associated
permissions. This is the same as saying the user has no permissions
for this library. Need to decide if we should allow the test that the
library is in fact existent before checking they are allowed to
access it.