Addresses Issue #10
Code implemented to include permissions when trying to add/remove content
to a library when someone has been given the correct permissions.

Relevant tests were added for views and webservices.

One interesting point raised was we still need to check the correct
content is sent to the api, and a sensible response. A nice pattern to
ensure no DRY.