Addresses and closes issue #7
Anonymous users will be handled by the API, such that they can only
access endpoints with no scope.

The GET endpoint for /libraries/<> was made scopeless.

As different methods cannot have their own scope, a separate endpoint
was created /documents/<>, to allow the POST and DELETE methods that
were originally a part of the libraries endpoint.

The tests were updated and split to encorporate this new endpoint,
and view.

A minor check that the correct view is scopeless was added to the
anonymous epic test.