Addresses and closes issue #24
Added a functional test to check that only the person with correct
permissions can delete a library.

Relevants tests were updated for the changes.

Stub data was updated to include the possibility to add names to
the User stub data, which can make it easier to identify any fails
in the test.