Issue #32 and #41
This closes issues 32 and 41.

A library is created with a default library name 'Untitled Library + <int:N>'.
A library is created with a default description 'My ADS library'.

There is now a PUT end point for the UserView, at /libraries/<string:library>.
The user must have the permissions of 'admin' or 'owner' to update the name
or the description.
The user can only update either 'name' or 'description'. If they pass any other
information, it will be ignored. The return contains the updated value.

Relevant functional, view, and webservices tests have been included.