Issue #36
Added:

Can include bibcode: [...] to POST when you create a library via the /libraries
end point.

Functionality test for creating a library and passing bibcodes on the creation.

Extend and shorten methods to MutableLists in the models. This means the lists
can be directly added to the current list, and all items within a list can be
removed from the existing list.

Lists are also made unique before being passed for creation. Duplication tests
were also added for the documents. Users should not be able to pass bibcodes that
are already in the list. It is probably worth checking if a unique constrain can
be placed on the content of a postgreSQL ARRAY.

Changed:

Add/remove of bibcodes now expect a bibcode with a type of list.

Tests updated to reflect the new type of bibcode.

Documentation updated for the methods and the '# TODO:' items moved out of the
class definition and left as comments within the class instead. This removes the
spam from the /resources end points.