Issue #30
The /libraries end point was updated to include:

num_users: the number of users that have permissions for a library.
Only the admin and owner can see this information, otherwise it
defaults to zero

num_documents: number of bibcodes in a library

date_created

date_last_modified

permission: finds the highest ranking permissions and returns that

Relevant tests have been included. Before this can be pushed, there
are two other steps that need to be taken:

 - migrations for the models must be included
 - deletion tests, that ensure that permissions are removed once
   the library has been deleted. This cascading should be working,
   but it is worth adding a test for it to be sure.
 - check views.py:L405. This can be improved and some tests added
   for it.