Issue #33: pipeline layed
Tests laid out that require a solr like output in the form of
functional tests, unit tests in the webservices, and a unit test
in the views.

Mocking is required and the beginning of that has been implemented
in base.py. However, to more correctly reflect the data required
a little bit more detail is required from the end point itself.

The configuration file was updated to include the end point of the
solr service. The 'api' scope is required to access this end point.
Before it is deployed to the AWS, the consul key store should be
updated.