Issue #33: solr proxy included and tests fleshed out
Mock service included for the solr end point. Currently, it only
returns data that is similar to that of the end point. However, if
there are issues in the future, something more realistic will have
to be incorporated, but for now it is nice to have the scaffolding.

Consul has been updated appropriately.