Feature, number of added/removed docs returned
The number of documents that are added or removed using the libraries/<>
end point with the POST method, are returned to the user with the json
keys 'number_added' or 'number_removed', depending if you used the action
'add' or 'remove', respectively.

Relevant tests have been udpated to include this in the response.