Issue #56 major additions towards closing
Test database added for the API. In memory is not used because this
would require two different scripts being able to access it. This is
viable using a shared cache approach, but would have to sit in the
main body of code that is not wanted. Instead, a small database is
created in the /tmp/ directory.

Tests have been included to check that the user, their libraries, and
all permissions get removed if they are deleted from the database,
as is expected.