Issue #76: closes and addresses cronjob required for production
Added a cronjob that can be run once every day at 1 am to remove stale users.

To keep the configuration as-is, a command line argument was allowed to be
passed to the Flask-Script Command class. This means on deployment, and in
testing, both versions of the script should behave as expected.

No tests added, just updated.

Minor spelling modification in views.py.