Issue #AWS-12
Started to turn the Array type into JSON.

Added some generic functions to add/remove libraries so that one does
not need to directly interface with the data type. This will allow
changes in the future to be easy.

Updated some of the tests to pass, but obviously, not everything is
passing yet, and needs to be checked.