Issue #AWS-12
Added a utils function that compares to lists that are unsorted. This
assumes you pass a hashable type, and that you can have repeating values
within the dictionary. In addition, a simple test wrapper for flask-test
was added in the base tests file.

Some stub data was updated as bibcodes were transitioned to a dict/json
type rather than a list.

The MutableList tests and its type were removed completely from the models
code, given MutableDict is being used instead.

Other tests were modified to start using the relevant add/remove bibcodes.
Currently, the tests for views, and models are passing, and so a nice place
to have a commit.