Issue AWS-12
Webservice level tests are now passing, the following explicit changes
were made in the following to achieve that:

A new util function that compares two unsorted lists was added. The
tests for utils was updated to include ones for this.

A new util function that returns the item of a dictionary, in a list
of dictionaries, given a key, without looping over the entire list,
was added. The relevant tests were also included.

The stub data was slightly modified to mimick the class object used
for the Library model.