Issue AWS-12
All the functional tests are now passing. The only modification was
updating the stub library access of bibcodes to get_bibcodes(), to
ensure that a list was returned. This is due to the fact that the stub
library now stores bibcodes as a dictionary (JSON) format, as does
the model.

This should mean there are no affects to the front-end service, given
the interface has not been altered in any way, in terms of the input
and output expected from each end point.

Any new problems can be raised in issues and easily fixed and implemented
in a matter of minutes (*touch wood*).