AWS-211 implement get classic libraries
Added an import end point for the classic service. This is a simple
get end point that starts the pulling and compiling of libraries from
ADS classic to BBB, assuming that the user has setup an ADS account
in their settings. If they have not, this will not do anything.

The process is idempotent. There are tests for two different scenarios,
that should cover most of the use cases.